refactor(gui): split Combos workspace for react-doctor
Extract ComboWorkspace panels/modals and harden fetch/loading paths for the 0.9.1 doctor rules.
