+++ gui/src/components/ComboWorkspace.tsx
@@ -1,838 +1,20 @@
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useCallback, useMemo, useState } from "react";
 import {
-  type ComboEffort,
   type ComboItem,
-  type ComboStrategy,
-  type ComboTarget,
-  COMBO_EFFORTS,
-  buildComboAttention,
   comboModelId,
-  comboPublicModelId,
-  draftEquals,
   emptyDraft,
   filterCombos,
   groupCombos,
-  validateComboDraft,
 } from "../combo-workspace-data";
-import {
-  IconAlert,
-  IconChevron,
-  IconGrip,
-  IconPlus,
-  IconSearch,
-  IconShuffle,
-  IconTrash,
-  IconX,
-} from "../icons";
-import { useT } from "../i18n";
-import { Notice } from "../ui";
+import { IconChevron, IconPlus, IconSearch, IconShuffle } from "../icons";
+import { useT } from "../i18n/shared";
+import { AddComboModal } from "./combo-workspace-add-modal";
+import { DetailPanel } from "./combo-workspace-detail-panel";
+import { RemoveComboDialog, UnsavedLeaveDialog } from "./combo-workspace-dialogs";
+import { OverviewPanel } from "./combo-workspace-overview-panel";
+import type { ComboWorkspaceProps } from "./combo-workspace-types";
 
-export type ProviderOption = {
-  name: string;
-  disabled?: boolean;
-  hiddenFromPicker?: boolean;
-  authMode?: string;
-  adapter?: string;
-  baseUrl?: string;
-};
-export type ModelOption = { provider: string; id: string; namespaced?: string };
-
-type DetailTab = "config" | "about";
-
-export interface ComboWorkspaceProps {
-  combos: ComboItem[];
-  providers: ProviderOption[];
-  models: ModelOption[];
-  loading?: boolean;
-  onRefresh: () => void;
-  onSave: (item: ComboItem, isCreate: boolean, renameFrom?: string) => Promise<{ ok: boolean; error?: string }>;
-  onRemove: (id: string) => Promise<{ ok: boolean; error?: string }>;
-  onAdd: () => void;
-  adding: boolean;
-  onCloseAdd: () => void;
-  onCreated: (id: string) => void;
-}
-
-function enabledProviders(providers: ProviderOption[]): ProviderOption[] {
-  return providers
-    .filter((p) => !p.disabled && !p.hiddenFromPicker)
-    .sort((a, b) => a.name.localeCompare(b.name));
-}
-
-/** ChatGPT passthrough has no /models catalog — GPT slugs are listed under provider "openai". */
-function isChatGptForwardOption(p: ProviderOption | undefined): boolean {
-  if (!p) return false;
-  const id = p.name.toLowerCase();
-  if (id !== "openai" && id !== "chatgpt") return false;
-  if ((p.authMode ?? "").toLowerCase() !== "forward") return false;
-  if ((p.adapter ?? "").toLowerCase() !== "openai-responses") return false;
-  const base = (p.baseUrl ?? "").replace(/\/+$/, "");
-  return !base || base.includes("chatgpt.com/backend-api/codex");
-}
-
-function modelsForProvider(
-  models: ModelOption[],
-  provider: string,
-  providers: ProviderOption[],
-): string[] {
-  const keys = new Set<string>([provider]);
-  const meta = providers.find((p) => p.name === provider);
-  // Alias chatgpt → openai native GPT rows (forward providers don't publish their own catalog).
-  if (provider.toLowerCase() === "chatgpt" || isChatGptForwardOption(meta)) {
-    keys.add("openai");
-  }
-  const ids = models
-    .filter((m) => keys.has(m.provider))
-    .map((m) => m.id)
-    .filter(Boolean);
-  return [...new Set(ids)].sort((a, b) => a.localeCompare(b));
-}
-
-function StrategySeg({
-  value,
-  onChange,
-  disabled,
-}: {
-  value: ComboStrategy;
-  onChange: (next: ComboStrategy) => void;
-  disabled?: boolean;
-}) {
-  const t = useT();
-  return (
-    <div className="cwi-strategy-seg" role="radiogroup" aria-label={t("cws.strategy")}>
-      {([
-        ["failover", "cws.strategy.failover"],
-        ["round-robin", "cws.strategy.roundRobin"],
-      ] as const).map(([id, key]) => (
-        <button
-          key={id}
-          type="button"
-          role="radio"
-          aria-checked={value === id}
-          className={`btn btn-sm${value === id ? " btn-primary" : " btn-ghost"}`}
-          disabled={disabled}
-          onClick={() => onChange(id)}
-        >
-          {t(key)}
-        </button>
-      ))}
-    </div>
-  );
-}
-
-function EffortSelect({
-  id,
-  value,
-  onChange,
-  disabled,
-}: {
-  id: string;
-  value: ComboEffort | null;
-  onChange: (next: ComboEffort | null) => void;
-  disabled?: boolean;
-}) {
-  const t = useT();
-  return (
-    <select
-      id={id}
-      className="input"
-      value={value ?? ""}
-      disabled={disabled}
-      aria-label={t("cws.field.defaultEffort")}
-      onChange={(e) => onChange(e.target.value === "" ? null : e.target.value as ComboEffort)}
-    >
-      <option value="">{t("cws.field.defaultEffortNone")}</option>
-      {COMBO_EFFORTS.map((effort) => (
-        <option key={effort} value={effort}>{effort}</option>
-      ))}
-    </select>
-  );
-}
-
-function TargetEditor({
-  targets,
-  strategy,
-  providers,
-  models,
-  onChange,
-}: {
-  targets: ComboTarget[];
-  strategy: ComboStrategy;
-  providers: ProviderOption[];
-  models: ModelOption[];
-  onChange: (next: ComboTarget[]) => void;
-}) {
-  const t = useT();
-  const provs = enabledProviders(providers);
-  const [dragIndex, setDragIndex] = useState<number | null>(null);
-  const [overIndex, setOverIndex] = useState<number | null>(null);
-
-  const update = (index: number, patch: Partial<ComboTarget>) => {
-    onChange(targets.map((row, i) => (i === index ? { ...row, ...patch } : row)));
-  };
-
-  const reorder = (from: number, to: number) => {
-    if (from === to || from < 0 || to < 0 || from >= targets.length || to >= targets.length) return;
-    const copy = [...targets];
-    const [moved] = copy.splice(from, 1);
-    copy.splice(to, 0, moved!);
-    onChange(copy);
-  };
-
-  return (
-    <div className="cwi-target-list">
-      {targets.map((row, index) => {
-        const currentProvider = providers.find((provider) => provider.name === row.provider);
-        const providerOptions = currentProvider && !provs.some((provider) => provider.name === row.provider)
-          ? [...provs, currentProvider]
-          : provs;
-        const modelIds = modelsForProvider(models, row.provider, providers);
-        const options = row.model && !modelIds.includes(row.model)
-          ? [row.model, ...modelIds]
-          : modelIds;
-        const modelSelectDisabled = !row.provider;
-        const dragging = dragIndex === index;
-        const dropTarget = overIndex === index && dragIndex !== null && dragIndex !== index;
-        return (
-          <div
-            key={index}
-            className={[
-              "cwi-target-row",
-              strategy === "failover" ? "cwi-target-row--failover" : "",
-              dragging ? "cwi-target-row--dragging" : "",
-              dropTarget ? "cwi-target-row--drop" : "",
-            ].filter(Boolean).join(" ")}
-            onDragOver={(e) => {
-              if (dragIndex === null) return;
-              e.preventDefault();
-              e.dataTransfer.dropEffect = "move";
-              if (overIndex !== index) setOverIndex(index);
-            }}
-            onDrop={(e) => {
-              e.preventDefault();
-              if (dragIndex !== null) reorder(dragIndex, index);
-              setDragIndex(null);
-              setOverIndex(null);
-            }}
-            onDragEnd={() => {
-              setDragIndex(null);
-              setOverIndex(null);
-            }}
-          >
-            <button
-              type="button"
-              className="cwi-target-grip"
-              draggable
-              aria-label={t("cws.target.drag")}
-              title={t("cws.target.drag")}
-              onDragStart={(e) => {
-                setDragIndex(index);
-                e.dataTransfer.effectAllowed = "move";
-                e.dataTransfer.setData("text/plain", String(index));
-              }}
-            >
-              <IconGrip width={14} height={14} aria-hidden="true" />
-            </button>
-            <select
-              className="input"
-              value={row.provider}
-              aria-label={t("cws.target.provider")}
-              onChange={(e) => {
-                const provider = e.target.value;
-                const first = modelsForProvider(models, provider, providers)[0] ?? "";
-                update(index, { provider, model: first });
-              }}
-            >
-              <option value="">{t("cws.target.pickProvider")}</option>
-              {providerOptions.map((p) => (
-                <option key={p.name} value={p.name}>
-                  {p.disabled ? t("cws.target.disabled", { name: p.name }) : p.name}
-                </option>
-              ))}
-            </select>
-            <select
-              className="input"
-              value={row.model}
-              disabled={modelSelectDisabled}
-              aria-label={t("cws.target.model")}
-              onChange={(e) => update(index, { model: e.target.value })}
-            >
-              <option value="">
-                {modelSelectDisabled
-                  ? t("cws.target.pickProviderFirst")
-                  : options.length === 0
-                    ? t("cws.target.noModels")
-                    : t("cws.target.pickModel")}
-              </option>
-              {options.map((id) => (
-                <option key={id} value={id}>{id}</option>
-              ))}
-            </select>
-            {strategy === "round-robin" && (
-              <input
-                className="input mono"
-                type="number"
-                min={1}
-                max={10000}
-                value={row.weight ?? 1}
-                aria-label={t("cws.target.weight")}
-                onChange={(e) => update(index, { weight: Number(e.target.value) })}
-              />
-            )}
-            <div className="cwi-target-actions">
-              <button
-                type="button"
-                className="btn btn-ghost btn-sm"
-                disabled={targets.length <= 1}
-                onClick={() => onChange(targets.filter((_, i) => i !== index))}
-                aria-label={t("common.remove")}
-              >
-                <IconTrash width={14} height={14} />
-              </button>
-            </div>
-          </div>
-        );
-      })}
-      <button
-        type="button"
-        className="btn btn-ghost btn-sm"
-        style={{ alignSelf: "flex-start" }}
-        onClick={() => onChange([...targets, { provider: "", model: "" }])}
-      >
-        <IconPlus width={14} height={14} /> {t("cws.target.add")}
-      </button>
-    </div>
-  );
-}
-
-function RemoveComboDialog({
-  model,
-  onCancel,
-  onConfirm,
-}: {
-  model: string;
-  onCancel: () => void;
-  onConfirm: () => void;
-}) {
-  const t = useT();
-  useEffect(() => {
-    const onKey = (e: KeyboardEvent) => {
-      if (e.key === "Escape") {
-        e.preventDefault();
-        onCancel();
-      }
-    };
-    window.addEventListener("keydown", onKey);
-    return () => window.removeEventListener("keydown", onKey);
-  }, [onCancel]);
-
-  return (
-    <div className="modal-overlay" role="dialog" aria-modal="true" aria-labelledby="cwi-remove-title" onClick={onCancel}>
-      <div className="modal-card pwi-remove-confirm-card" onClick={(e) => e.stopPropagation()}>
-        <h3 id="cwi-remove-title" className="pwi-remove-confirm-title">
-          {t("cws.removeConfirmTitle", { model })}
-        </h3>
-        <p className="muted pwi-remove-confirm-desc">{t("cws.removeConfirmDesc")}</p>
-        <div className="pwi-remove-confirm-actions">
-          <button type="button" className="btn btn-ghost" onClick={onCancel}>{t("common.cancel")}</button>
-          <button type="button" className="btn pwi-remove-confirm-danger" onClick={onConfirm}>{t("common.remove")}</button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function UnsavedLeaveDialog({
-  onKeep,
-  onDiscard,
-}: {
-  onKeep: () => void;
-  onDiscard: () => void;
-}) {
-  const t = useT();
-  return (
-    <div className="modal-overlay" role="dialog" aria-modal="true" aria-labelledby="cwi-unsaved-title" onClick={onKeep}>
-      <div className="modal-card pwi-json-unsaved-card" onClick={(e) => e.stopPropagation()}>
-        <h3 id="cwi-unsaved-title" className="pwi-json-unsaved-title">{t("cws.unsavedTitle")}</h3>
-        <p className="muted pwi-json-unsaved-desc">{t("cws.unsavedDesc")}</p>
-        <div className="pwi-json-unsaved-actions">
-          <button type="button" className="btn btn-ghost" onClick={onKeep}>{t("cws.keepEditing")}</button>
-          <button type="button" className="btn btn-danger" onClick={onDiscard}>{t("common.discard")}</button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-export function AddComboModal({
-  existingIds,
-  existingAliases,
-  providerMap,
-  providers,
-  models,
-  onClose,
-  onSubmit,
-}: {
-  existingIds: string[];
-  existingAliases: string[];
-  providerMap: Readonly<Record<string, { disabled?: boolean }>>;
-  providers: ProviderOption[];
-  models: ModelOption[];
-  onClose: () => void;
-  onSubmit: (item: ComboItem) => Promise<{ ok: boolean; error?: string }>;
-}) {
-  const t = useT();
-  const [draft, setDraft] = useState<ComboItem>(() => emptyDraft());
-  const [busy, setBusy] = useState(false);
-  const [error, setError] = useState("");
-
-  useEffect(() => {
-    const onKey = (e: KeyboardEvent) => {
-      if (e.key === "Escape" && !busy) {
-        e.preventDefault();
-        onClose();
-      }
-    };
-    window.addEventListener("keydown", onKey);
-    return () => window.removeEventListener("keydown", onKey);
-  }, [busy, onClose]);
-
-  const submit = async () => {
-    const code = validateComboDraft(draft, {
-      existingIds,
-      existingAliases,
-      isCreate: true,
-      providers: providerMap,
-    });
-    if (code) {
-      setError(t(`cws.err.${code}`));
-      return;
-    }
-    setBusy(true);
-    setError("");
-    const id = draft.id.trim();
-    const alias = draft.alias?.trim() || null;
-    const res = await onSubmit({ ...draft, id, alias, model: comboPublicModelId(id, alias) });
-    setBusy(false);
-    if (!res.ok) {
-      setError(res.error || t("cws.saveFailed"));
-      return;
-    }
-  };
-
-  return (
-    <div className="modal-overlay" role="dialog" aria-modal="true" aria-labelledby="cwi-add-title" onClick={() => { if (!busy) onClose(); }}>
-      <div className="modal-card" style={{ width: "min(560px, 94vw)" }} onClick={(e) => e.stopPropagation()}>
-        <div className="row" style={{ justifyContent: "space-between", marginBottom: 8 }}>
-          <h3 id="cwi-add-title" style={{ margin: 0 }}>{t("cws.addTitle")}</h3>
-          <button type="button" className="btn btn-ghost btn-sm" onClick={onClose} disabled={busy} aria-label={t("common.close")}>
-            <IconX width={16} height={16} />
-          </button>
-        </div>
-        <p className="muted" style={{ marginTop: 0 }}>{t("cws.addSubtitle")}</p>
-        {error && <Notice tone="err">{error}</Notice>}
-        <div className="cwi-modal-form">
-          <div className="cwi-field">
-            <label htmlFor="cwi-new-id">{t("cws.field.id")}</label>
-            <input
-              id="cwi-new-id"
-              className="input mono"
-              value={draft.id}
-              disabled={busy}
-              onChange={(e) => setDraft((d) => ({
-                ...d,
-                id: e.target.value,
-                model: comboPublicModelId(e.target.value, d.alias),
-              }))}
-            />
-            <p className="muted" style={{ fontSize: 12, margin: "4px 0 0" }}>
-              {t("cws.field.idInternalHint")}
-            </p>
-          </div>
-          <div className="cwi-field">
-            <label htmlFor="cwi-new-alias">{t("cws.field.alias")}</label>
-            <input
-              id="cwi-new-alias"
-              className="input mono"
-              value={draft.alias ?? ""}
-              placeholder={t("cws.field.aliasPlaceholder")}
-              disabled={busy}
-              onChange={(e) => setDraft((d) => ({
-                ...d,
-                alias: e.target.value.trim() ? e.target.value : null,
-                model: comboPublicModelId(d.id, e.target.value),
-              }))}
-            />
-            <p className="muted" style={{ fontSize: 12, margin: "4px 0 0" }}>
-              {t("cws.field.aliasHint")}
-            </p>
-            <p className="muted" style={{ fontSize: 12, margin: "4px 0 0" }}>
-              {t("cws.field.idHint", {
-                model: draft.id.trim() ? comboPublicModelId(draft.id, draft.alias) : "…",
-              })}
-            </p>
-          </div>
-          <div className="cwi-field">
-            <label>{t("cws.strategy")}</label>
-            <StrategySeg
-              value={draft.strategy}
-              disabled={busy}
-              onChange={(strategy) => setDraft((d) => ({ ...d, strategy }))}
-            />
-            <p className="muted" style={{ fontSize: 12, margin: "6px 0 0" }}>
-              {draft.strategy === "failover" ? t("cws.strategy.failoverHint") : t("cws.strategy.roundRobinHint")}
-            </p>
-          </div>
-          <div className="cwi-field">
-            <label htmlFor="cwi-new-effort">{t("cws.field.defaultEffort")}</label>
-            <EffortSelect
-              id="cwi-new-effort"
-              value={draft.defaultEffort}
-              disabled={busy}
-              onChange={(defaultEffort) => setDraft((d) => ({ ...d, defaultEffort }))}
-            />
-            <p className="muted" style={{ fontSize: 12, margin: "4px 0 0" }}>
-              {t("cws.field.defaultEffortHint")}
-            </p>
-          </div>
-          {draft.strategy === "round-robin" && (
-            <div className="cwi-field">
-              <label htmlFor="cwi-new-sticky">{t("cws.field.stickyLimit")}</label>
-              <input
-                id="cwi-new-sticky"
-                className="input mono"
-                type="number"
-                min={1}
-                max={100}
-                value={draft.stickyLimit}
-                disabled={busy}
-                onChange={(e) => setDraft((d) => ({ ...d, stickyLimit: Number(e.target.value) }))}
-              />
-              <p className="muted" style={{ fontSize: 12, margin: "4px 0 0" }}>
-                {t("cws.field.stickyLimitHint")}
-              </p>
-            </div>
-          )}
-          <div className="cwi-field">
-            <label>{t("cws.targets")}</label>
-            <p className="muted" style={{ fontSize: 12, margin: "0 0 8px" }}>
-              {draft.strategy === "failover" ? t("cws.targets.failoverHint") : t("cws.targets.roundRobinHint")}
-            </p>
-            <TargetEditor
-              targets={draft.targets}
-              strategy={draft.strategy}
-              providers={providers}
-              models={models}
-              onChange={(targets) => setDraft((d) => ({ ...d, targets }))}
-            />
-          </div>
-        </div>
-        <div className="cwi-modal-actions">
-          <button type="button" className="btn btn-ghost" onClick={onClose} disabled={busy}>{t("common.cancel")}</button>
-          <button type="button" className="btn btn-primary" onClick={() => { void submit(); }} disabled={busy}>
-            {busy ? t("common.saving") : t("cws.create")}
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function OverviewPanel({
-  combos,
-  onSelect,
-  onAdd,
-}: {
-  combos: ComboItem[];
-  onSelect: (id: string) => void;
-  onAdd: () => void;
-}) {
-  const t = useT();
-  const sections = groupCombos(combos);
-  const attention = buildComboAttention(combos);
-
-  return (
-    <div className="combos-workspace-overview">
-      <div className="combos-workspace-overview-head">
-        <h2 className="combos-workspace-overview-title">{t("cws.overviewTitle")}</h2>
-        <button type="button" className="btn btn-primary btn-sm" onClick={onAdd}>
-          <IconPlus width={14} height={14} /> {t("cws.add")}
-        </button>
-      </div>
-      <p className="muted" style={{ marginTop: 0, maxWidth: "62ch" }}>{t("cws.overviewBlurb")}</p>
-      <div className="cwi-count-strip">
-        <div className="cwi-count-pill"><strong>{combos.length}</strong><span>{t("cws.count.total")}</span></div>
-        <div className="cwi-count-pill"><strong>{sections.failover.length}</strong><span>{t("cws.count.failover")}</span></div>
-        <div className="cwi-count-pill"><strong>{sections.roundRobin.length}</strong><span>{t("cws.count.roundRobin")}</span></div>
-      </div>
-
-      <section className="pwi-section" aria-label={t("cws.howTitle")}>
-        <h3 className="pwi-section-title">{t("cws.howTitle")}</h3>
-        <p className="muted" style={{ margin: 0 }}>{t("cws.howBody")}</p>
-      </section>
-
-      {attention.length > 0 && (
-        <section className="pwi-section" aria-label={t("cws.attentionTitle")}>
-          <h3 className="pwi-section-title">{t("cws.attentionTitle")}</h3>
-          <div className="cwi-attention-list">
-            {attention.map((item) => (
-              <button key={item.id} type="button" className="cwi-attention-row" onClick={() => onSelect(item.id)}>
-                <IconAlert width={14} height={14} aria-hidden="true" />
-                <code className="chip">{item.model}</code>
-                <span className="muted">
-                  {item.reason === "empty-targets" ? t("cws.attention.empty") : t("cws.attention.few")}
-                </span>
-                <IconChevron width={14} height={14} style={{ marginLeft: "auto" }} aria-hidden="true" />
-              </button>
-            ))}
-          </div>
-        </section>
-      )}
-    </div>
-  );
-}
-
-function DetailPanel({
-  baseline,
-  isCreate = false,
-  otherIds,
-  otherAliases,
-  providerMap,
-  providers,
-  models,
-  onBack,
-  onSaved,
-  onRequestRemove,
-  onSave,
-  onDirtyChange,
-}: {
-  baseline: ComboItem;
-  isCreate?: boolean;
-  /** Ids of all OTHER combos — rename collisions validate against these. */
-  otherIds: string[];
-  /** Aliases of all OTHER combos — alias uniqueness validates against these. */
-  otherAliases: string[];
-  providerMap: Readonly<Record<string, { disabled?: boolean }>>;
-  providers: ProviderOption[];
-  models: ModelOption[];
-  onBack?: () => void;
-  onSaved: (item: ComboItem) => void;
-  onRequestRemove?: () => void;
-  onSave: (item: ComboItem, isCreate: boolean, renameFrom?: string) => Promise<{ ok: boolean; error?: string }>;
-  onDirtyChange: (dirty: boolean) => void;
-}) {
-  const t = useT();
-  const [tab, setTab] = useState<DetailTab>("config");
-  const [draft, setDraft] = useState<ComboItem>(baseline);
-  const [busy, setBusy] = useState(false);
-  const [msg, setMsg] = useState<{ ok: boolean; text: string } | null>(null);
-  const [copied, setCopied] = useState(false);
-  const dirty = !draftEquals(draft, baseline);
-  const baselineSyncKey = `${baseline.id}:${baseline.alias ?? ""}:${baseline.strategy}:${baseline.stickyLimit}:${baseline.defaultEffort}:${baseline.targets.map((t) => `${t.provider}/${t.model}:${t.weight ?? 1}`).join(",")}`;
-
-  useEffect(() => {
-    const timer = window.setTimeout(() => {
-      setDraft(baseline);
-      setMsg(null);
-      setTab("config");
-    }, 0);
-    return () => window.clearTimeout(timer);
-    // Sync when server/local baseline content changes for this combo (same id after save/refresh).
-    // eslint-disable-next-line react-hooks/exhaustive-deps -- intentional: key captures baseline payload
-  }, [baselineSyncKey]);
-
-  useEffect(() => {
-    onDirtyChange(dirty);
-  }, [dirty, onDirtyChange]);
-
-  const copyModel = async () => {
-    try {
-      await navigator.clipboard.writeText(baseline.model);
-      setCopied(true);
-      window.setTimeout(() => setCopied(false), 1200);
-    } catch {
-      /* ignore */
-    }
-  };
-
-  const save = async () => {
-    const code = validateComboDraft(draft, {
-      existingIds: otherIds,
-      existingAliases: otherAliases,
-      isCreate,
-      providers: providerMap,
-    });
-    if (code) {
-      setMsg({ ok: false, text: t(`cws.err.${code}`) });
-      return;
-    }
-    setBusy(true);
-    const trimmedId = draft.id.trim();
-    const alias = draft.alias?.trim() || null;
-    const item = {
-      ...draft,
-      id: trimmedId,
-      alias,
-      model: comboPublicModelId(trimmedId, alias),
-    };
-    const renameFrom = !isCreate && trimmedId !== baseline.id ? baseline.id : undefined;
-    const res = await onSave(item, isCreate, renameFrom);
-    setBusy(false);
-    if (!res.ok) {
-      setMsg({ ok: false, text: res.error || t("cws.saveFailed") });
-      return;
-    }
-    setMsg({
-      ok: true,
-      text: isCreate ? t("cws.created", { model: item.model }) : t("cws.saved"),
-    });
-    onSaved(item);
-  };
-
-  const headerModel = isCreate
-    ? (draft.id.trim() ? comboPublicModelId(draft.id, draft.alias) : t("cws.addTitle"))
-    : baseline.model;
-
-  return (
-    <div className="combos-workspace-detail">
-      <div className="combos-workspace-detail-head">
-        {onBack && (
-          <button type="button" className="btn btn-ghost btn-sm pwi-back-overview" onClick={onBack} aria-label={t("cws.backToAll")}>
-            <IconChevron style={{ width: 14, height: 14, transform: "rotate(180deg)" }} aria-hidden="true" />
-            {t("cws.allCombos")}
-          </button>
-        )}
-        <h2 className="combos-workspace-detail-title">{headerModel}</h2>
-        {!isCreate && (
-          <button type="button" className="chip cwi-copy-chip" onClick={() => { void copyModel(); }} title={t("cws.copyModel")}>
-            {copied ? t("cws.copied") : t("cws.copyModel")}
-          </button>
-        )}
-        <div className="combos-workspace-detail-actions">
-          {!isCreate && onRequestRemove && (
-            <button type="button" className="btn btn-ghost btn-sm" onClick={onRequestRemove}>
-              <IconTrash width={14} height={14} /> {t("common.remove")}
-            </button>
-          )}
-          <button type="button" className="btn btn-primary btn-sm" disabled={(!isCreate && !dirty) || busy} onClick={() => { void save(); }}>
-            {busy ? t("common.saving") : t(isCreate ? "cws.create" : "common.save")}
-          </button>
-        </div>
-      </div>
-
-      {msg && <Notice tone={msg.ok ? "ok" : "err"}>{msg.text}</Notice>}
-
-      <div className="combos-workspace-tabs" role="tablist">
-        <button type="button" role="tab" aria-selected={tab === "config"} className={`combos-workspace-tab${tab === "config" ? " combos-workspace-tab--active" : ""}`} onClick={() => setTab("config")}>
-          {t("cws.tab.config")}
-        </button>
-        <button type="button" role="tab" aria-selected={tab === "about"} className={`combos-workspace-tab${tab === "about" ? " combos-workspace-tab--active" : ""}`} onClick={() => setTab("about")}>
-          {t("cws.tab.about")}
-        </button>
-      </div>
-
-      <div className="combos-workspace-tab-content" role="tabpanel">
-        {tab === "config" ? (
-          <div className="cwi-form-grid">
-            <div className="cwi-field">
-              <label htmlFor="cwi-edit-id">{t("cws.field.id")}</label>
-              <input
-                id="cwi-edit-id"
-                className="input mono"
-                value={draft.id}
-                disabled={busy}
-                onChange={(e) => setDraft((d) => ({
-                  ...d,
-                  id: e.target.value,
-                  model: comboPublicModelId(e.target.value, d.alias),
-                }))}
-              />
-              <p className="muted" style={{ fontSize: 12, margin: "4px 0 0" }}>
-                {isCreate
-                  ? t("cws.field.idInternalHint")
-                  : t("cws.field.idHintEdit", { model: comboPublicModelId(draft.id, draft.alias) })}
-              </p>
-            </div>
-            <div className="cwi-field">
-              <label htmlFor="cwi-edit-alias">{t("cws.field.alias")}</label>
-              <input
-                id="cwi-edit-alias"
-                className="input mono"
-                value={draft.alias ?? ""}
-                placeholder={comboModelId(draft.id.trim() || "…")}
-                disabled={busy}
-                onChange={(e) => setDraft((d) => ({
-                  ...d,
-                  alias: e.target.value.trim() ? e.target.value : null,
-                  model: comboPublicModelId(d.id, e.target.value),
-                }))}
-              />
-              <p className="muted" style={{ fontSize: 12, margin: "4px 0 0" }}>
-                {t("cws.field.aliasHint")}
-              </p>
-            </div>
-            <div className="cwi-field">
-              <label>{t("cws.strategy")}</label>
-              <StrategySeg
-                value={draft.strategy}
-                disabled={busy}
-                onChange={(strategy) => setDraft((d) => ({ ...d, strategy }))}
-              />
-              <p className="muted" style={{ fontSize: 12, margin: "6px 0 0" }}>
-                {draft.strategy === "failover" ? t("cws.strategy.failoverHint") : t("cws.strategy.roundRobinHint")}
-              </p>
-            </div>
-            <div className="cwi-field">
-              <label htmlFor="cwi-effort">{t("cws.field.defaultEffort")}</label>
-              <EffortSelect
-                id="cwi-effort"
-                value={draft.defaultEffort}
-                disabled={busy}
-                onChange={(defaultEffort) => setDraft((d) => ({ ...d, defaultEffort }))}
-              />
-              <p className="muted" style={{ fontSize: 12, margin: "4px 0 0" }}>
-                {t("cws.field.defaultEffortHint")}
-              </p>
-            </div>
-            {draft.strategy === "round-robin" && (
-              <div className="cwi-field">
-                <label htmlFor="cwi-sticky">{t("cws.field.stickyLimit")}</label>
-                <input
-                  id="cwi-sticky"
-                  className="input mono"
-                  type="number"
-                  min={1}
-                  max={100}
-                  value={draft.stickyLimit}
-                  disabled={busy}
-                  onChange={(e) => setDraft((d) => ({ ...d, stickyLimit: Number(e.target.value) }))}
-                />
-              </div>
-            )}
-            <div className="cwi-field">
-              <label>{t("cws.targets")}</label>
-              <p className="muted" style={{ fontSize: 12, margin: "0 0 8px" }}>
-                {draft.strategy === "failover" ? t("cws.targets.failoverHint") : t("cws.targets.roundRobinHint")}
-              </p>
-              <TargetEditor
-                targets={draft.targets}
-                strategy={draft.strategy}
-                providers={providers}
-                models={models}
-                onChange={(targets) => setDraft((d) => ({ ...d, targets }))}
-              />
-            </div>
-          </div>
-        ) : (
-          <section className="pwi-section">
-            <h3 className="pwi-section-title">{t("cws.aboutTitle")}</h3>
-            <p className="muted" style={{ margin: 0 }}>{t("cws.aboutBody")}</p>
-          </section>
-        )}
-      </div>
-    </div>
-  );
-}
+export type { ModelOption, ProviderOption, ComboWorkspaceProps } from "./combo-workspace-types";
 
 export default function ComboWorkspace({
   combos,
